move blocklist event to subfolder
